:recycle: Toast の実装から循環参照を除去する
