+++ apps/routing/src/02-nest-routes-deep/pages/Friends/index.tsx
@@ -3,6 +3,7 @@ import { gutter } from '@learn-react/core/src/helpers/Style';
 import { generatePath, Navigate, NavLink, Route, Routes } from 'react-router-dom';
 import { Router } from '../../constants/Router';
 import { Friend } from './Friend';
+import { FriendData } from './data';
 
 export const Friends = () => (
   <main className={styleBase}>
@@ -50,35 +51,3 @@ const List = () => (
     ))}
   </ul>
 );
-
-type FriendType = {
-  id: string;
-  nameJa: string;
-  nameEn: string;
-  family: string;
-};
-
-const FriendData: FriendType[] = [
-  {
-    id: 'serval',
-    nameJa: 'サーバル',
-    nameEn: 'Serval Cat',
-    family: 'ネコ目ネコ科ネコ属',
-  },
-  {
-    id: 'raccoon',
-    nameJa: 'アライグマ',
-    nameEn: 'Common raccoon',
-    family: 'ネコ目アライグマ科アライグマ属',
-  },
-  {
-    id: 'fennec',
-    nameJa: 'フェネック',
-    nameEn: 'Fennec',
-    family: 'ネコ目イヌ科キツネ属',
-  },
-];
-
-export function getFriendById(id: string): FriendType | undefined {
-  return FriendData.find((f) => f.id === id);
-}
